+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -1,7 +1,7 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.UserHistory;
-import nextstep.jdbc.JdbcTemplate;
+import nextstep.jdbc.core.JdbcTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
