+++ jdbc/src/main/java/nextstep/jdbc/support/StatementCallback.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.support;
 
 import java.sql.SQLException;
 import java.sql.Statement;
